+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### 🐞 Fixed
 - Reactions picker for large messages sometimes goes in the safe area
+- Loading of pinned messages in the channel info screen
 
 # [4.50.1](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.50.1)
 _March 14, 2024_